Move AxisHelper to components directory and fix imports
