+++ src/LineDrawer.tsx
@@ -14,8 +14,8 @@ import {
   Entity,
   HeightReference,
 } from 'cesium'
-import AxisHelper from './AxisHelper'
-import type { AxisHelperProps } from './AxisHelper'
+import AxisHelper from './components/axis-helper'
+import type { AxisHelperProps } from './components/axis-helper'
 
 interface LineDrawerProps {
   viewer: Viewer | null
